piedpiper-noop-executor/requirements.txt
PyYAML
requests
flake8
minio
piedpiperlib@git+https://github.com/ragingpastry/python-piedpiper.git@artman-libs